+++ Source/Bogus/Extensions/ExtensionsForString.cs
@@ -1,3 +1,6 @@
+using System.Globalization;
+using System.Text;
+
 namespace Bogus.Extensions
 {
     /// <summary>
@@ -27,5 +30,27 @@ namespace Bogus.Extensions
             }
             return str;
         }
+
+       /// <summary>
+       /// A string extension method that removes the diacritics character from the strings.
+       /// </summary>
+       /// <param name="this">The @this to act on.</param>
+       /// <returns>The string without diacritics character.</returns>
+       public static string RemoveDiacritics(this string @this)
+       {
+          string normalizedString = @this.Normalize(NormalizationForm.FormD);
+          var sb = new StringBuilder();
+
+          foreach( char t in normalizedString )
+          {
+             UnicodeCategory uc = CharUnicodeInfo.GetUnicodeCategory(t);
+             if( uc != UnicodeCategory.NonSpacingMark )
+             {
+                sb.Append(t);
+             }
+          }
+
+          return sb.ToString().Normalize(NormalizationForm.FormC);
+       }
     }
 }